// src/schemas/contact-schema.ts
import { z } from "zod";

export const contactSchema = z.object({
  name: z.string().min(2, "Name must be at least 2 characters"),

  email: z.string().email("Invalid email address"),

  subject: z.string().min(3, "Subject is required"),

  message: z.string().min(10, "Message must be at least 10 characters"),

  website: z.string().optional(),
});

export type ContactSchemaType = z.infer<typeof contactSchema>;